requirements.txt
aiohttp>=3.8.0
Pillow>=9.0.0
asyncio>=3.4.3
python-dotenv>=0.19.0